refactor(bolt): move bytecode plan renderer
Move Stage 6 bytecode read-RAF Rust constant rendering into the narrowly named plan module.

This leaves the Stage 6 emitter responsible only for splicing the rendered constants and keeps generated verifier output unchanged.
