+++ crates/bolt/src/protocols/jolt/emit/rust/stage6.rs
@@ -22,10 +22,7 @@ use super::output_claims::{
 };
 use crate::emit::rust::{push_format, EmitError, RustSourceFile};
 use crate::ir::{string_attribute_value, symbol_attribute_value, BoltModule, Cpu, Role};
-use crate::protocols::jolt::stage6_bytecode_read_raf_plan::{
-    stage6_bytecode_read_raf_plan, BytecodeFlag, BytecodeReadRafPlan, BytecodeReadRafTermPlan,
-    BytecodeRegister,
-};
+use crate::protocols::jolt::stage6_bytecode_read_raf_plan::emit_stage6_bytecode_read_raf_plan_constants;
 use crate::schema::verify_cpu_schema;
 
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -1466,9 +1463,7 @@ const STAGE6_RELATION_SYMBOLS: Stage67RelationSymbols = Stage67RelationSymbols {
 };
 "#,
         );
-        source.push_str(&emit_stage6_bytecode_read_raf_plan(
-            stage6_bytecode_read_raf_plan(),
-        ));
+        source.push_str(&emit_stage6_bytecode_read_raf_plan_constants());
         source.push_str(
             r#"
 #[derive(Debug)]
@@ -2665,153 +2660,6 @@ fn intern_str_array(
     source.push_str(&emit_str_array(&name, values));
     arrays.push((values.to_vec(), name.clone()));
     name
-}
-
-fn emit_stage6_bytecode_read_raf_plan(plan: &BytecodeReadRafPlan) -> String {
-    let mut source = "\n".to_owned();
-
-    for stage in plan.stages {
-        push_format(
-            &mut source,
-            format_args!(
-                "const {}: &[Stage67BytecodeTermPlan] = &[\n",
-                stage.terms_const
-            ),
-        );
-        for term in stage.terms {
-            push_format(
-                &mut source,
-                format_args!("    {},\n", emit_stage6_bytecode_term_plan(term)),
-            );
-        }
-        source.push_str("];\n");
-    }
-
-    push_format(
-        &mut source,
-        format_args!(
-            "const {}: &[Stage67BytecodeStagePlan] = &[\n",
-            plan.stages_const
-        ),
-    );
-    for stage in plan.stages {
-        push_format(
-            &mut source,
-            format_args!(
-                "    Stage67BytecodeStagePlan {{ gamma: {}, cycle_point: {}, register_point: {}, output_gamma_power: {}, identity_gamma_power: {}, terms: {} }},\n",
-                rust_str(stage.gamma),
-                rust_str(stage.cycle_point),
-                rust_option_str(stage.register_point),
-                stage.output_gamma_power,
-                emit_option_usize(stage.identity_gamma_power),
-                stage.terms_const,
-            ),
-        );
-    }
-    source.push_str("];\n\n");
-
-    push_format(
-        &mut source,
-        format_args!(
-            "const {}: Stage67BytecodeReadRafPlan = Stage67BytecodeReadRafPlan {{\n",
-            plan.const_name
-        ),
-    );
-    push_format(
-        &mut source,
-        format_args!(
-            "    point: {},\n    gamma: {},\n    bytecode_ra_eval_prefix: {},\n    entries: {},\n    entry_bytecode_index: {},\n    stages: {},\n",
-            rust_str(plan.point),
-            rust_str(plan.gamma),
-            rust_str(plan.bytecode_ra_eval_prefix),
-            rust_str(plan.entries),
-            rust_str(plan.entry_bytecode_index),
-            plan.stages_const,
-        ),
-    );
-    push_format(
-        &mut source,
-        format_args!(
-            "    entry_contribution: Stage67BytecodeEntryContributionPlan {{ gamma_power: {} }},\n",
-            plan.entry_contribution.gamma_power
-        ),
-    );
-    source.push_str("    registers: Stage67BytecodeRegisterSymbols {\n");
-    push_format(
-        &mut source,
-        format_args!(
-            "        rd: {},\n        rs1: {},\n        rs2: {},\n",
-            rust_str(plan.registers.rd),
-            rust_str(plan.registers.rs1),
-            rust_str(plan.registers.rs2),
-        ),
-    );
-    source.push_str("    },\n");
-    push_format(
-        &mut source,
-        format_args!(
-            "    entry_lookup_table: {},\n",
-            rust_str(plan.entry_lookup_table)
-        ),
-    );
-    source.push_str("};\n");
-    source
-}
-
-fn emit_stage6_bytecode_term_plan(term: &BytecodeReadRafTermPlan) -> String {
-    match term {
-        BytecodeReadRafTermPlan::Address { gamma_power } => {
-            format!("Stage67BytecodeTermPlan::Address {{ gamma_power: {gamma_power} }}")
-        }
-        BytecodeReadRafTermPlan::Imm { gamma_power } => {
-            format!("Stage67BytecodeTermPlan::Imm {{ gamma_power: {gamma_power} }}")
-        }
-        BytecodeReadRafTermPlan::CircuitFlag { index, gamma_power } => format!(
-            "Stage67BytecodeTermPlan::CircuitFlag {{ index: {index}, gamma_power: {gamma_power} }}"
-        ),
-        BytecodeReadRafTermPlan::EntryFlag {
-            flag,
-            expected,
-            gamma_power,
-        } => format!(
-            "Stage67BytecodeTermPlan::EntryFlag {{ flag: {}, expected: {expected}, gamma_power: {gamma_power} }}",
-            emit_stage6_bytecode_flag(*flag)
-        ),
-        BytecodeReadRafTermPlan::RegisterEq {
-            register,
-            gamma_power,
-        } => format!(
-            "Stage67BytecodeTermPlan::RegisterEq {{ register: {}, gamma_power: {gamma_power} }}",
-            emit_stage6_bytecode_register(*register)
-        ),
-        BytecodeReadRafTermPlan::LookupTable { gamma_base } => {
-            format!("Stage67BytecodeTermPlan::LookupTable {{ gamma_base: {gamma_base} }}")
-        }
-    }
-}
-
-fn emit_stage6_bytecode_flag(flag: BytecodeFlag) -> &'static str {
-    match flag {
-        BytecodeFlag::IsInterleaved => "Stage67BytecodeFlag::IsInterleaved",
-        BytecodeFlag::IsBranch => "Stage67BytecodeFlag::IsBranch",
-        BytecodeFlag::LeftIsRs1 => "Stage67BytecodeFlag::LeftIsRs1",
-        BytecodeFlag::LeftIsPc => "Stage67BytecodeFlag::LeftIsPc",
-        BytecodeFlag::RightIsRs2 => "Stage67BytecodeFlag::RightIsRs2",
-        BytecodeFlag::RightIsImm => "Stage67BytecodeFlag::RightIsImm",
-        BytecodeFlag::IsNoop => "Stage67BytecodeFlag::IsNoop",
-    }
-}
-
-fn emit_stage6_bytecode_register(register: BytecodeRegister) -> &'static str {
-    match register {
-        BytecodeRegister::Rd => "Stage67BytecodeRegister::Rd",
-        BytecodeRegister::Rs1 => "Stage67BytecodeRegister::Rs1",
-        BytecodeRegister::Rs2 => "Stage67BytecodeRegister::Rs2",
-    }
-}
-
-fn emit_option_usize(value: Option<usize>) -> String {
-    value.map_or_else(|| "None".to_owned(), |value| format!("Some({value})"))
 }
 
 fn rust_str(value: &str) -> String {
